move tests and pyarrow imports
